client/rpc: Move Transport to its own module
